moved css from styling into the right folders and files and then deleted it
