+++ FrontEnd/src/components/HeaderNavComponent/HeaderNavComponent.css
@@ -1,0 +1,24 @@
+.header-nav-bar{
+    display: flex;
+    justify-content: space-between;
+    height: 4em;
+    background-color: black;
+    padding: .5em;
+}
+.logo-img{
+    margin: 0.5em;
+    height: 2.5em;
+}
+.hmbrg-img{
+    margin: 0.5em;
+    margin-right: 1em;
+    height: 2.5em;
+}
+@media (min-width: 768px){
+    .header-nav-bar{
+        background-color: rgba(0, 0, 0, 0);
+    }
+    .hmbrg-img{
+        display: none;
+    }
+}
