+++ FrontEnd/src/components/HeaderNavComponent/HeaderNavComponent.jsx
@@ -1,9 +1,9 @@
 import { Link } from "react-router-dom";
 import { useState, useEffect } from "react";
-import logo from "../assets/FilmvisarnaLogoTwo.png";
-import menu from "../assets/menu.png";
-import NavComponent from "./NavComponent.jsx";
-import "../styling.css"
+import logo from "../../assets/FilmvisarnaLogoTwo.png";
+import menu from "../../assets/menu.png";
+import NavComponent from "../NavComponent/NavComponent.jsx";
+import "./HeaderNavComponent.css"
 
 export default function HeaderNavComponent() {
   const [isDesktop, setIsDesktop] = useState(window.innerWidth >= 768);
