+++ FrontEnd/src/components/NavComponent/NavComponent.css
@@ -1,0 +1,114 @@
+.nav-container{
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    width: 100%;
+    height: 100vh;
+    background: linear-gradient(#150316 70%, #8027CE 150%);
+    position: absolute;
+    text-align: center;
+    top: 0em;
+    left: 0em;
+    z-index: 1000;
+}
+.logo-img-nav{
+    height: 2.5em;
+    top: 1em;
+    position: absolute;
+    left: 1em;
+}
+.hmbrg-img-nav{
+    position: absolute;
+    height: 1.5em;
+    right: 0.5em;
+    margin: 1.5em;
+}
+.nav-div-container{
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    height: 90%;
+}
+.nav-list-container{
+    list-style-type: none;
+    margin: 0;
+    padding: 0; 
+}
+.nav-list-item{
+    margin: auto;
+    padding: 1em 0;
+    color:#C699EA;
+    width: 100%;
+    text-transform: capitalize;
+    font-size: 1.2em;
+    font-weight: 700;
+}
+.active{
+    background-image: linear-gradient(to right, #BA7E36 2%, #8027CE 2%);
+    text-shadow: #150316;
+}
+.nav-login-btn{
+    width: 80%;
+    margin: 4em 0;
+    background-color: #C699EA;
+    color: #150316;
+}
+@media (min-width: 768px){
+    .hmbrg-img-nav{
+        display: none;
+    }
+    .nav-container{
+        flex-direction: row;
+        height: 4em;
+        /* From https://css.glass */
+        background: rgba(44, 20, 46, 0.18);
+        box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
+        backdrop-filter: blur(2.6px);
+        border: 1px solid rgba(44, 20, 46, 0.29);
+    }
+    .nav-div-container{
+        display: flex;
+        flex-direction: row;
+        justify-content: end;
+        width: 70%;
+        align-items: baseline;
+    }
+    .nav-list-container{
+        display: flex;
+        flex-direction: row;
+        justify-content: space-evenly;
+        width: 100%;
+    }
+    .nav-list-item{
+        padding: 1em 0;
+        width: 100%;
+        font-size: 1em;
+        font-weight: 500;
+    }
+    .active{
+        background-image: linear-gradient(rgba(0, 0, 0, 0), rgba(0, 0, 0, 0));
+        color:#BA7E36;
+    }
+    .nav-login-btn{
+        width: 80%;
+        margin-bottom: 0em;
+        margin: 0;
+        background-color: black;
+        color:#C699EA;
+        border-color: #C699EA;
+        height: 2em;
+        width: 7em;
+        padding: 0;
+        margin-right: 1em;
+    }
+}
+@media (min-width: 1500px){
+    .nav-div-container{
+        width: 50%;
+    }
+}
+@media (min-width: 2000px){
+    .nav-div-container{
+        width: 30%;
+    }
+}
