--- FrontEnd/src/styling.css
@@ -1,310 +1,0 @@
-.main-text-title, .date-title, .redan-medlem{
-    color: #BA7E36;
-}
-.main-text-title{
-    text-align: center;
-    width: 350px;
-    margin: 20px auto;
-    font-weight: 100;
-}
-.date-title{
-    margin-left: 5%;
-}
-.iframe-container, .button-container, .text-containers, iframe{
-    margin: auto;
-    width: 350px;
-}
-.button-container{
-    margin-bottom: 1em;
-}
-iframe, .iframe-container {
-    height: 200px;
-}
-body{
-    background: radial-gradient(circle at center, rgb(70, 10, 73), black);
-    width: 100%;
-    background-attachment: fixed;
-}
-.movieinfo-container{
-    margin: auto;
-}
-.main-btn{
-    width: 100%;
-    background-color: #C699EA;
-    color: #150316;
-}
-.main-btn-color{
-    margin: auto;
-    max-height: 50px;
-    background-color: #C699EA;
-    color: #150316;
-}
-.sec-btn{
-    width: 100%;
-    margin-top: 20px;
-    background-color: #8027CE;
-    color: #C699EA;
-}
-.seperator{
-    margin: auto;
-    height: 2px;
-    width: 95%;
-    background-color: #BA7E36;
-}
-.screenings-list-container{
-    list-style-type: none;
-    margin: 0;
-    padding: 0;
-    background: rgba(0, 0, 0, 0.18);
-    border-radius: 20px;
-}
-.screenings-list{
-    margin: 5%;
-    display: flex;
-}
-.list-movie-title, .screenings-p{
-    width: 25%;
-    margin: 0px auto;
-}
-.link-color{
-    color:rgb(204, 211, 218);
-}
-.screenings-selectors{
-    width: 23.5%;
-    background-color: #D9D9D9;
-    border-radius: 5px;
-    color: black;
-    font-size: large;
-    margin: auto;
-}
-.filters-containers{
-    display: flex;
-    flex-direction: row;
-    margin-bottom: 1em;
-}
-div#root{
-    width: 100%;
-}
-.normal-text, .screenings-p, .link-color{
-    color:#BA7E36;
-}
-
-/* HeaderNavComponent css --------------------------------------------------------------------------------------------------------------------*/
-.header-nav-bar{
-    display: flex;
-    justify-content: space-between;
-    height: 4em;
-    background-color: black;
-    padding: .5em;
-}
-.logo-img{
-    margin: 0.5em;
-    height: 2.5em;
-}
-.hmbrg-img{
-    margin: 0.5em;
-    margin-right: 1em;
-    height: 2.5em;
-}
-/* NavComponent css  --------------------------------------------------------------------------------------------------------------------*/
-.logo-img-nav{
-    height: 2.5em;
-    top: 1em;
-    position: absolute;
-    left: 1em;
-}
-.hmbrg-img-nav{
-    position: absolute;
-    height: 1.5em;
-    right: 0.5em;
-    margin: 1.5em;
-}
-.nav-container{
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    width: 100%;
-    height: 100vh;
-    background: linear-gradient(#150316 70%, #8027CE 150%);
-    position: absolute;
-    text-align: center;
-    top: 0em;
-    left: 0em;
-    z-index: 1000;
-}
-.nav-list-container{
-    list-style-type: none;
-    margin: 0;
-    padding: 0; 
-}
-.nav-list-item{
-    margin: auto;
-    padding: 1em 0;
-    color:#C699EA;
-    width: 100%;
-    text-transform: capitalize;
-    font-size: 1.2em;
-    font-weight: 700;
-}
-.nav-login-btn{
-    width: 80%;
-    margin: 4em 0;
-    background-color: #C699EA;
-    color: #150316;
-}
-.active{
-    background-image: linear-gradient(to right, #BA7E36 2%, #8027CE 2%);
-    text-shadow: #150316;
-}
-.nav-div-container{
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    height: 90%;
-}
-/*register login page  ------------------------------------------------------------------------------------------*/
-/* .form-container{
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    width: 350px;
-    margin: auto;
-}
-.register-lbl, .login-lbl{
-    color:#C699EA;
-    font-weight: 500;
-    margin-left: 1em;
-}
-.input-container{
-    display: flex;
-    flex-direction: column;
-}
-.register-field, .login-field{
-    background-color: #2C142E;
-    border: 2px solid #150316;
-    width: 100%;
-    height: 3em;
-    padding-left: 1em;
-    box-sizing: border-box;
-    border-radius: 1em;
-    margin: auto;
-    outline: none;
-}
-.register-field:focus, .login-field:focus{
-    border: 2px solid #BA7E36;
-}
-.page-wrapper{
-    height: 100vh;
-}
-.welcome-text{
-    font-weight: 300;
-    font-size: 2.5rem;
-    width: 350px;
-    margin: auto;
-    text-align: center;
-} */
-
-@media (max-width: 350px){
-    .iframe-container, .button-container, .text-containers, iframe, .main-text-title{
-        margin: auto;
-        width: 250px;
-    }
-    iframe, .iframe-container {
-        height: 141px;
-    }
-}
-@media (min-width: 600px){
-    .iframe-container, iframe{
-        width: 500px;
-        height: 281.25px;
-    }
-    .text-containers, .button-container, .main-text-title{
-        width: 500px;
-    }
-    .screenings-selectors{
-        width: 23.5%;
-    }
-    /*login register*/
-    /* .welcome-text, .form-container{
-        width: 450px;
-    } */
-}
-@media (min-width: 768px){
-    /* header nav component  --------------------------------------------------------------------------------------------------------------------*/
-    .hmbrg-img, .hmbrg-img-nav{
-        display: none;
-    }
-    .header-nav-bar{
-        background-color: rgba(0, 0, 0, 0);
-    }
-    .nav-list-container{
-        display: flex;
-        flex-direction: row;
-        justify-content: space-evenly;
-        width: 100%;
-    }
-    .nav-container{
-        flex-direction: row;
-        height: 4em;
-        /* From https://css.glass */
-        background: rgba(44, 20, 46, 0.18);
-        box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
-        backdrop-filter: blur(2.6px);
-        border: 1px solid rgba(44, 20, 46, 0.29);
-    }
-    .nav-login-btn{
-        width: 80%;
-        margin-bottom: 0em;
-    }
-    .nav-list-item{
-        padding: 1em 0;
-        width: 100%;
-        font-size: 1em;
-        font-weight: 500;
-    }
-    .nav-div-container{
-        display: flex;
-        flex-direction: row;
-        justify-content: end;
-        width: 70%;
-        align-items: baseline;
-    }
-    .nav-login-btn{
-        margin: 0;
-        background-color: black;
-        color:#C699EA;
-        border-color: #C699EA;
-        height: 2em;
-        width: 7em;
-        padding: 0;
-        margin-right: 1em;
-    }
-    .active{
-        background-image: linear-gradient(rgba(0, 0, 0, 0), rgba(0, 0, 0, 0));
-        color:#BA7E36;
-    }
-    /*more info page  --------------------------------------------------------------------------------------------------------------------*/
-    .iframe-container, iframe{
-        width: 700px;
-        height: 393.75px;
-    }
-    .text-containers, .button-container, .main-text-title{
-        width: 700px;
-    }
-    .return-container{
-        margin: auto;
-        width: 750px;
-    }
-}
-@media (min-width: 1500px){
-    /* header nav component  --------------------------------------------------------------------------------------------------------------------*/
-    .nav-div-container{
-        width: 50%;
-    }
-}
-@media (min-width: 2000px){
-     /* header nav component  --------------------------------------------------------------------------------------------------------------------*/
-    .nav-div-container{
-        width: 30%;
-    }
-}
